# Conflicts:
#   CONTRIBUTING.md

+++ src/cook/cook_build.rs
@@ -1,6 +1,7 @@
-use pkg::package::PackageError;
+use pkg::PackageError;
 use pkg::{Package, PackageName};
 
+use crate::config::CookConfig;
 use crate::cook::fs::*;
 use crate::cook::package::{package_source_paths, package_target};
 use crate::cook::pty::PtyOut;
@@ -172,15 +173,19 @@ pub fn build(
     target_dir: &Path,
     name: &PackageName,
     recipe: &Recipe,
-    offline_mode: bool,
+    cook_config: &CookConfig,
     check_source: bool,
     logger: &PtyOut,
 ) -> Result<(Vec<PathBuf>, BTreeSet<PackageName>), String> {
     let sysroot_dir = target_dir.join("sysroot");
     let toolchain_dir = target_dir.join("toolchain");
     let stage_dirs = get_stage_dirs(&recipe.optional_packages, target_dir);
-    let cli_verbose = crate::config::get_config().cook.verbose;
-    let cli_jobs = crate::config::get_config().cook.jobs;
+    let stage_pkgars: Vec<PathBuf> = stage_dirs
+        .iter()
+        .map(|p| p.with_added_extension("pkgar"))
+        .collect();
+    let cli_verbose = cook_config.verbose;
+    let cli_jobs = cook_config.jobs;
     if recipe.build.kind == BuildKind::None {
         // metapackages don't need to do anything here
         return Ok((stage_dirs, BTreeSet::new()));
@@ -193,8 +198,8 @@ pub fn build(
         &recipe.build.dev_dependencies[..],
     ]
     .concat();
-    let build_deps = CookRecipe::get_build_deps_recursive(&build_deps, false, false)
-        .map_err(|e| format!("{:?}", e))?;
+    let build_deps =
+        CookRecipe::get_build_deps_recursive(&build_deps, false).map_err(|e| format!("{:?}", e))?;
     for dependency in build_deps.iter() {
         let (_, pkgar, _) = dependency.stage_paths();
         if dependency.name.is_host() {
@@ -204,9 +209,32 @@ pub fn build(
         }
     }
 
-    if !check_source && stage_dirs.iter().all(|dir| dir.exists()) {
-        let auto_deps = build_auto_deps(recipe, target_dir, &stage_dirs, dep_pkgars, logger)?;
-        return Ok((stage_dirs, auto_deps));
+    macro_rules! make_auto_deps {
+        () => {
+            build_auto_deps(
+                recipe,
+                target_dir,
+                &stage_dirs,
+                cook_config,
+                dep_pkgars,
+                logger,
+            )
+        };
+    }
+
+    if !check_source {
+        let stage_present = if cook_config.clean_target {
+            stage_pkgars.iter().all(|file| file.is_file())
+        } else {
+            stage_dirs.iter().all(|file| file.is_dir())
+        };
+        if stage_present {
+            if cli_verbose {
+                log_to_pty!(logger, "DEBUG: using cached build, not checking source");
+            }
+            let auto_deps = make_auto_deps!()?;
+            return Ok((stage_dirs, auto_deps));
+        }
     }
 
     let mut source_modified = modified_dir_ignore_git(source_dir).unwrap_or(SystemTime::UNIX_EPOCH);
@@ -215,6 +243,7 @@ pub fn build(
             source_modified = recipe_modified
         }
     }
+
     let deps_modified = dep_pkgars
         .iter()
         .map(|(_dep, pkgar)| modified(pkgar))
@@ -226,9 +255,33 @@ pub fn build(
         .max()
         .unwrap_or(Ok(SystemTime::UNIX_EPOCH))?;
 
+    // check stage dir modified against pkgar files, any files missing will result in UNIX_EPOCH
+    let stage_modified = modified_all(&stage_pkgars, modified).unwrap_or(SystemTime::UNIX_EPOCH);
+    // Rebuild stage if source is newer
+    if stage_modified < source_modified
+        || stage_modified < deps_modified
+        || stage_modified < deps_host_modified
+    {
+        for stage_dir in &stage_dirs {
+            if stage_dir.is_dir() {
+                log_to_pty!(logger, "DEBUG: updating '{}'", stage_dir.display());
+                remove_stage_dir(stage_dir)?;
+            }
+        }
+    } else {
+        if cli_verbose {
+            log_to_pty!(logger, "DEBUG: using cached build");
+        }
+        if cook_config.clean_target {
+            // stop early otherwise we'll end up rebuilding
+            let auto_deps = make_auto_deps!()?;
+            return Ok((stage_dirs, auto_deps));
+        }
+    }
+
     // Rebuild sysroot if source is newer
     if recipe.build.kind != BuildKind::Remote {
-        build_deps_dir(
+        let updated = build_deps_dir(
             logger,
             &sysroot_dir,
             target_dir.join("sysroot.tmp"),
@@ -240,9 +293,12 @@ pub fn build(
             source_modified,
             deps_modified,
         )?;
+        if cli_verbose && !updated {
+            log_to_pty!(logger, "DEBUG: using cached sysroot");
+        }
     }
     if recipe.build.kind != BuildKind::Remote && !name.is_host() && dep_host_pkgars.len() > 0 {
-        build_deps_dir(
+        let updated = build_deps_dir(
             logger,
             &toolchain_dir,
             target_dir.join("toolchain.tmp"),
@@ -250,35 +306,22 @@ pub fn build(
             source_modified,
             deps_host_modified,
         )?;
-    }
-
-    // Rebuild stage if source is newer
-    if stage_dirs.iter().any(|dir| dir.is_dir()) {
-        let stage_modified =
-            modified_all(&stage_dirs, modified_dir).unwrap_or(SystemTime::UNIX_EPOCH);
-        if stage_modified < source_modified
-            || stage_modified < deps_modified
-            || stage_modified < deps_host_modified
-        {
-            for stage_dir in &stage_dirs {
-                log_to_pty!(logger, "DEBUG: updating '{}'", stage_dir.display());
-                remove_stage_dir(stage_dir)?;
-            }
+        if cli_verbose && !updated {
+            log_to_pty!(logger, "DEBUG: using cached toolchain");
         }
     }
 
-    if !stage_dirs.last().is_some_and(|dir| dir.is_dir()) {
-        let stage_dir = stage_dirs
-            .last()
-            .expect("Should have atleast one stage dir");
+    let stage_dir = stage_dirs
+        .last()
+        .expect("Should have atleast one stage dir");
+    let build_dir = get_build_dir(target_dir);
+    if !stage_dir.is_dir() {
         // Create stage.tmp
         let stage_dir_tmp = target_dir.join("stage.tmp");
         create_dir_clean(&stage_dir_tmp)?;
 
-        // Create build, if it does not exist
-        //TODO: flag for clean builds where build is wiped out
-        let build_dir = target_dir.join("build");
-        if !build_dir.is_dir() {
+        // Create build dir, if it does not exist
+        if cook_config.clean_build || !build_dir.is_dir() {
             create_dir_clean(&build_dir)?;
         }
 
@@ -294,20 +337,46 @@ pub fn build(
         };
 
         if recipe.build.kind == BuildKind::Remote {
-            return build_remote(stage_dirs, recipe, target_dir);
+            return build_remote(stage_dirs, recipe, target_dir, cook_config);
         }
+
         //TODO: better integration with redoxer (library instead of binary)
         //TODO: configurable target
         //TODO: Add more configurability, convert scripts to Rust?
         let script = match &recipe.build.kind {
             BuildKind::Cargo {
-                package_path,
+                cargopath,
                 cargoflags,
+                cargopackages,
+                cargoexamples,
             } => {
-                format!(
-                    "DYNAMIC_INIT\nPACKAGE_PATH={} cookbook_cargo {cargoflags}",
-                    package_path.as_deref().unwrap_or(".")
-                )
+                let mut script = format!(
+                    "DYNAMIC_INIT\n{}\nCOOKBOOK_CARGO_PATH={} ",
+                    flags_fn("COOKBOOK_CARGO_FLAGS", cargoflags),
+                    cargopath.as_deref().unwrap_or(".")
+                );
+                if cargopackages.len() == 0 && cargoexamples.len() == 0 {
+                    script += "cookbook_cargo\n"
+                } else {
+                    if cargopackages.len() > 0 {
+                        script += "cookbook_cargo_packages";
+                        for package in cargopackages {
+                            script += " ";
+                            script += package;
+                        }
+                        script += "\n";
+                    }
+                    if cargoexamples.len() > 0 {
+                        script += "cookbook_cargo_examples";
+                        for example in cargoexamples {
+                            script += " ";
+                            script += example;
+                        }
+                        script += "\n";
+                    }
+                }
+
+                script
             }
             BuildKind::Configure { configureflags } => format!(
                 "DYNAMIC_INIT\n{}cookbook_configure",
@@ -336,13 +405,14 @@ pub fn build(
             let cookbook_sysroot = sysroot_dir.canonicalize().unwrap();
             let cookbook_toolchain = toolchain_dir.canonicalize().ok();
             let bash_args = if cli_verbose { "-ex" } else { "-e" };
-            let mut command = if is_redox() {
-                let mut command = Command::new("bash");
+            let local_redoxer = Path::new("target/release/cookbook_redoxer");
+            let mut command = if is_redox() && !local_redoxer.is_file() {
+                let mut command = Command::new("cookbook_redoxer");
                 command.arg(bash_args);
-                command.env("COOKBOOK_REDOXER", "cargo");
+                command.env("COOKBOOK_REDOXER", "cookbook_redoxer");
                 command
             } else {
-                let cookbook_redoxer = Path::new("target/release/cookbook_redoxer")
+                let cookbook_redoxer = local_redoxer
                     .canonicalize()
                     .unwrap_or(PathBuf::from("/bin/false"));
                 let mut command = Command::new(&cookbook_redoxer);
@@ -369,7 +439,7 @@ pub fn build(
             if cli_verbose {
                 command.env("COOKBOOK_VERBOSE", "1");
             }
-            if offline_mode {
+            if cook_config.offline {
                 command.env("COOKBOOK_OFFLINE", "1");
             }
             command
@@ -408,8 +478,16 @@ pub fn build(
         rename(&stage_dir_tmp, &stage_dir)?;
     }
 
-    let auto_deps = build_auto_deps(recipe, target_dir, &stage_dirs, dep_pkgars, logger)?;
+    if cook_config.clean_target {
+        remove_all(&build_dir)?;
+        remove_all(&sysroot_dir)?;
+        if toolchain_dir.is_dir() {
+            remove_all(&toolchain_dir)?;
+        }
+        // don't remove stage dir yet
+    }
 
+    let auto_deps = make_auto_deps!()?;
     Ok((stage_dirs, auto_deps))
 }
 
@@ -422,19 +500,42 @@ pub fn remove_stage_dir(stage_dir: &PathBuf) -> Result<(), String> {
         remove_all(&stage_file)?;
     }
     let stage_meta = stage_dir.with_added_extension("toml");
-    Ok(if stage_meta.is_file() {
+    if stage_meta.is_file() {
         remove_all(&stage_meta)?;
-    })
+    }
+    let stage_files = stage_dir.with_added_extension("files");
+    if stage_files.is_file() {
+        remove_all(&stage_files)?;
+    }
+    Ok(())
 }
 
 pub fn get_stage_dirs(features: &Vec<OptionalPackageRecipe>, target_dir: &Path) -> Vec<PathBuf> {
+    let mut target_dir = target_dir.to_path_buf();
+    if let Some(cross_target) = std::env::var("COOKBOOK_CROSS_TARGET").ok() {
+        if cross_target != "" {
+            // TODO: automatically pass COOKBOOK_CROSS_GNU_TARGET?
+            target_dir = target_dir.join(cross_target)
+        }
+    }
     let mut v = Vec::new();
     for f in features {
         v.push(target_dir.join(format!("stage.{}", f.name)));
     }
     // intentionally added last as it contains leftover files from package features
-    v.push(target_dir.join(format!("stage")));
+    v.push(target_dir.join("stage"));
     v
+}
+
+pub fn get_build_dir(target_dir: &Path) -> PathBuf {
+    let mut target_dir = target_dir.to_path_buf();
+    if let Some(cross_target) = std::env::var("COOKBOOK_CROSS_TARGET").ok() {
+        if cross_target != "" {
+            // TODO: automatically pass COOKBOOK_CROSS_GNU_TARGET?
+            target_dir = target_dir.join(cross_target)
+        }
+    }
+    target_dir.join("build")
 }
 
 fn build_deps_dir(
@@ -444,15 +545,19 @@ fn build_deps_dir(
     dep_pkgars: &BTreeSet<(PackageName, PathBuf)>,
     source_modified: SystemTime,
     deps_modified: SystemTime,
-) -> Result<(), String> {
+) -> Result<bool, String> {
     if deps_dir.is_dir() {
         let tags_dir = deps_dir.join(".tags");
         let sysroot_modified = modified_dir(&tags_dir).unwrap_or(SystemTime::UNIX_EPOCH);
         if sysroot_modified < source_modified
             || sysroot_modified < deps_modified
-            || dep_pkgars
-                .iter()
-                .any(|(pkg, _)| !tags_dir.join(pkg.as_str()).is_file())
+            || !check_files_present(
+                &tags_dir,
+                &dep_pkgars
+                    .iter()
+                    .map(|(name, _)| name.without_host())
+                    .collect(),
+            )?
         {
             log_to_pty!(logger, "DEBUG: updating '{}'", deps_dir.display());
             remove_all(deps_dir)?;
@@ -476,7 +581,7 @@ fn build_deps_dir(
 
         let pkey_path = "build/id_ed25519.pub.toml";
         for (name, archive_path) in dep_pkgars {
-            let tag_file = tags_dir.join(name.without_host().as_str());
+            let tag_file = tags_dir.join(name.without_prefix());
             fs::write(&tag_file, "")
                 .map_err(|e| format!("failed to write tag file {}: {:?}", tag_file.display(), e))?;
             pkgar::extract(pkey_path, &archive_path, deps_dir_tmp.to_str().unwrap()).map_err(
@@ -493,9 +598,11 @@ fn build_deps_dir(
 
         // Move sysroot.tmp to sysroot atomically
         rename(&deps_dir_tmp, deps_dir)?;
+
+        return Ok(true);
     }
 
-    Ok(())
+    Ok(false)
 }
 
 /// Calculate automatic dependencies
@@ -503,13 +610,22 @@ fn build_auto_deps(
     recipe: &Recipe,
     target_dir: &Path,
     stage_dirs: &Vec<PathBuf>,
+    cook_config: &CookConfig,
     mut dep_pkgars: BTreeSet<(PackageName, PathBuf)>,
     logger: &PtyOut,
 ) -> Result<BTreeSet<PackageName>, String> {
     let auto_deps_path = target_dir.join("auto_deps.toml");
-    if auto_deps_path.is_file() && modified(&auto_deps_path)? < modified_all(stage_dirs, modified)?
-    {
-        remove_all(&auto_deps_path)?
+    if auto_deps_path.is_file() && !cook_config.clean_target {
+        if modified(&auto_deps_path)? < modified_all(stage_dirs, modified)? {
+            if cook_config.verbose {
+                log_to_pty!(logger, "DEBUG: updating {}", auto_deps_path.display());
+            }
+            remove_all(&auto_deps_path)?;
+        } else {
+            if cook_config.verbose {
+                log_to_pty!(logger, "DEBUG: Using cached auto deps");
+            }
+        }
     }
 
     let auto_deps = if auto_deps_path.exists() {
@@ -539,14 +655,19 @@ pub fn build_remote(
     stage_dirs: Vec<PathBuf>,
     recipe: &Recipe,
     target_dir: &Path,
+    cook_config: &CookConfig,
 ) -> Result<(Vec<PathBuf>, BTreeSet<PackageName>), String> {
     let source_toml = target_dir.join("source.toml");
-    let source_pubkey = target_dir.join("id_ed25519.pub.toml");
+    let source_pubkey = "build/remotes/pub_key_static.redox-os.org.toml";
 
     let packages = recipe.get_packages_list();
     for (i, package) in packages.into_iter().enumerate() {
         // declare pkg dependencies as autodeps dependency
         let stage_dir = &stage_dirs[i];
+
+        if cook_config.clean_target && stage_dir.with_added_extension("pkgar").is_file() {
+            continue;
+        }
 
         if !stage_dir.is_dir() {
             let (_, source_pkgar, _) = package_source_paths(package, &target_dir);
@@ -565,9 +686,10 @@ pub fn build_remote(
     }
 
     let auto_deps_path = target_dir.join("auto_deps.toml");
-    if auto_deps_path.is_file() && modified(&auto_deps_path)? < modified_all(&stage_dirs, modified)?
-    {
-        remove_all(&auto_deps_path)?
+    if auto_deps_path.is_file() && !cook_config.clean_target {
+        if modified(&auto_deps_path)? < modified_all(&stage_dirs, modified)? {
+            remove_all(&auto_deps_path)?
+        }
     }
 
     let auto_deps = if auto_deps_path.exists() {
